refactor(core): improve base code
