+++ app/src/main/assets/about.properties
@@ -1,6 +1,6 @@
 about.version=
-about.build=3642
-about.date=jeu. oct. 24 14:02:15 2019
+about.build=3655
+about.date=jeu. oct. 24 16:22:30 2019
 about.commit=
 about.commitFull=
 about.github=https://github.com/flyve-mdm/flyve-mdm-android-agent
