+++ app/src/mqtt/java/org/flyve/mdm/agent/core/mqtt/MqttController.java
@@ -154,107 +154,6 @@ public class MqttController {
     }
 
 
-    /**
-     * Lock Device
-     * Example {"lock":"now|unlock"}
-     */
-    public void lockDevice(Boolean lock) {
-
-        if(MDMAgent.isSecureVersion()) {
-            return;
-        }
-
-        try {
-
-
-            if(lock) {
-                androidPolicies.lockScreen(LockActivity.class);
-                androidPolicies.lockDevice();
-                broadcastReceivedLog(MQTT_SEND, "Lock", "Device Lock");
-            } else {
-                Helpers.sendBroadcast("unlock", "org.flyvemdm.finishlock", context);
-            }
-        } catch (Exception ex) {
-            broadcastReceivedLog(ERROR, "Error on lockDevice", ex.getMessage());
-            FlyveLog.e(this.getClass().getName() + ", lockDevice", ex.getCause().getMessage());
-        }
-    }
-
-    /**
-     * Erase all device data include SDCard
-     */
-    public void wipe() {
-        if(MDMAgent.isSecureVersion()) {
-            return;
-        }
-
-        try {
-            androidPolicies.wipe();
-            broadcastReceivedLog(MQTT_SEND, "Wipe", "Wipe success");
-        } catch (Exception ex) {
-            FlyveLog.e(this.getClass().getName() + ", wipe", ex.getMessage());
-            broadcastReceivedLog(ERROR, "Error on wipe", ex.getMessage());
-        }
-    }
-
-    /**
-     * Unenroll the device
-     */
-    public boolean unenroll() {
-        if(MDMAgent.isSecureVersion()) {
-            return false;
-        }
-
-        // Send message with unenroll
-        String topic = mTopic + "/Status/Unenroll";
-        String payload = "{\"unenroll\": \"unenrolled\"}";
-        byte[] encodedPayload = new byte[0];
-        try {
-            encodedPayload = payload.getBytes(UTF_8);
-            MqttMessage message = new MqttMessage(encodedPayload);
-            client.publish(topic, message);
-            broadcastReceivedLog(MQTT_SEND, "Unenroll", "Unenroll success");
-
-            // clear cache
-            mqttData.deleteAll();
-
-            // Remove all the information
-            new ApplicationData(context).deleteAll();
-            new FileData(context).deleteAll();
-            new MqttData(context).deleteAll();
-            new PoliciesData(context).deleteAll();
-
-            // send message
-            Helpers.sendBroadcast(Helpers.broadCastMessage("action", "open", "splash"), Helpers.BROADCAST_MSG, this.context);
-
-            // show offline
-            Helpers.sendBroadcast(false, Helpers.BROADCAST_STATUS, this.context);
-
-            return true;
-        } catch (Exception ex) {
-            FlyveLog.e(this.getClass().getName() + ", unenroll", ex.getMessage());
-            broadcastReceivedLog(ERROR, "Error on unenroll", ex.getMessage());
-            return false;
-        }
-    }
-
-    /**
-     * Send the Status version of the agent
-     * payload: {"online": true}
-     */
-    public void sendOnlineStatus(Boolean status) {
-        String topic = mTopic + "/Status/Online";
-        String payload = "{\"online\": " + Boolean.toString( status ) + "}";
-        byte[] encodedPayload = new byte[0];
-        try {
-            encodedPayload = payload.getBytes(UTF_8);
-            MqttMessage message = new MqttMessage(encodedPayload);
-            IMqttDeliveryToken token = client.publish(topic, message);
-            broadcastReceivedLog(MQTT_SEND, "Send Online Status", "ID: " + token.getMessageId());
-        } catch (Exception ex) {
-            broadcastReceivedLog(ERROR, "Error on sendStatusVersion", ex.getMessage());
-        }
-    }
 
 
     /**
